Move SpectrumDatasetMaker
