+++ gammapy/makers/__init__.py
@@ -1,2 +1,3 @@
 from .background import *
 from .safe import *
+from .spectrum import *
