+++ gammapy/makers/background/tests/test_reflected.py
@@ -15,8 +15,8 @@ from gammapy.maps import WcsGeom, WcsNDMap
 from gammapy.makers import (
     ReflectedRegionsBackgroundMaker,
     ReflectedRegionsFinder,
+    SpectrumDatasetMaker
 )
-from gammapy.spectrum.make import SpectrumDatasetMaker
 from gammapy.datasets import SpectrumDataset
 from gammapy.utils.regions import compound_region_to_list
 from gammapy.utils.testing import (
