+++ gammapy/spectrum/__init__.py
@@ -1,5 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """1D spectrum analysis."""
 from .flux_point import *
-from .make import *
 from .sensitivity import *
